fix: removed react native from project and added plain react
