+++ backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "symbiond-backend",
-  "version": "0.4.0",
+  "version": "0.6.0",
   "description": "A portfolio project.",
   "main": "index.js",
   "type": "module",
