+++ docker-compose.yaml
@@ -6,7 +6,7 @@ services:
       - ./pgdata:/var/lib/postgresql/data
       - ./backend/db-init:/docker-entrypoint-initdb.d
     env_file:
-      - ./env/backend.env
+      - .env
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U $$POSTGRES_USER -d $$POSTGRES_DB"]
       interval: 5s
@@ -22,7 +22,7 @@ services:
       - ./backend:/app
       - /app/node_modules
     env_file:
-      - ./env/backend.env
+      - .env
     environment:
       - NODE_ENV=development
     depends_on:
