--- frontend/.env.example
@@ -1,1 +1,0 @@
-EXPO_PUBLIC_SERVER_IP=your-server-ipv4-address
