--- frontend/context/WebSocketContext.js
@@ -1,42 +1,0 @@
-import { createContext, useContext, useState, useEffect, useCallback } from 'react';
-import { io } from 'socket.io-client';
-
-const WebSocketContext = createContext(null);
-
-export const WebSocketProvider = ({ children }) => {
-  const [socket, setSocket] = useState(null);
-
-  useEffect(() => {
-    const newSocket = io(`http://${process.env.EXPO_PUBLIC_SERVER_IP}:3000/`);
-    setSocket(newSocket);
-
-    return () => {
-      newSocket.disconnect();
-    };
-  }, []);
-
-  const emitEvent = useCallback((event, data) => {
-    if (socket) {
-      socket.emit(event, data);
-    }
-  }, [socket]);
-
-  const value = {
-    socket,
-    emitEvent
-  };
-
-  return (
-    <WebSocketContext.Provider value={value}>
-      {children}
-    </WebSocketContext.Provider>
-  );
-};
-
-export const useWebSocket = () => {
-  const context = useContext(WebSocketContext);
-  if (context === null) {
-    throw new Error('useWebSocket must be used within a WebSocketProvider');
-  }
-  return context;
-};
