--- frontend/locales/en/calendar.json
@@ -1,4 +1,0 @@
-{
-  "tabTitle": "Calendar",
-  "headerTitle": "Calendar"
-}
