--- frontend/locales/en/common.json
@@ -1,4 +1,0 @@
-{
-  "greeting": "Hello, {{name}}",
-  "welcome": "Welcome to Symbiond!"
-}
