--- frontend/locales/en/index.js
@@ -1,5 +1,0 @@
-// locales/en/index.js
-export * as common from './common.json';
-export * as time from './time.json';
-export * as calendar from './calendar.json';
-export * as map from './map.json';
