--- frontend/locales/en/map.json
@@ -1,9 +1,0 @@
-{
-  "tabTitle": "Map",
-  "headerTitle": "Map",
-  "states": {
-    "Diamora": {
-      "capital": "Emerald Hollow"
-    }
-  }
-}
