--- frontend/locales/en/time.json
@@ -1,4 +1,0 @@
-{
-  "formattedDate": "{{month}} {{day}}, {{year}} A.F.",
-  "formattedTime": "{{time}}"
-}
