+++ frontend/package.json
@@ -1,48 +1,35 @@
 {
-  "name": "symbiond-frontend",
-  "version": "0.4.0",
-  "description": "A portfolio project.",
-  "author": "Dylan Bowman dbowman@symbiond.com",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "start": "expo start",
-    "test": "jest",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web"
-  },
+  "name": "frontend",
+  "version": "0.6.0",
+  "private": true,
   "dependencies": {
-    "@expo/metro-runtime": "^5.0.4",
-    "@react-native-community/slider": "4.5.6",
-    "@react-navigation/bottom-tabs": "^7.2.0",
-    "@react-navigation/native": "^7.0.14",
-    "@react-navigation/native-stack": "^7.2.0",
-    "expo": "~53.0.9",
-    "expo-asset": "~11.1.5",
-    "expo-file-system": "~18.1.10",
-    "expo-leaflet": "^1.0.3",
-    "expo-linear-gradient": "~14.1.4",
-    "expo-localization": "~16.1.5",
-    "expo-status-bar": "~2.2.3",
-    "i18n-js": "^4.5.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-native": "0.79.2",
-    "react-native-linear-gradient": "^2.8.3",
-    "react-native-pager-view": "6.7.1",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.10.0",
-    "react-native-web": "^0.20.0",
-    "react-native-webview": "^13.13.5",
-    "socket.io-client": "^4.8.1"
+    "@vis.gl/react-maplibre": "^8.0.4",
+    "maplibre-gl": "^5.7.3",
+    "react": "16.11.0",
+    "react-dom": "16.11.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
-  "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@testing-library/react-native": "^12.9.0",
-    "@types/react": "~19.0.10",
-    "jest": "^29.7.0",
-    "jest-expo": "~53.0.5",
-    "react-native-dotenv": "^3.4.11"
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
