--- frontend/screens/CalendarScreen.js
@@ -1,12 +1,0 @@
-import { View } from "react-native";
-import Calendar from "../components/Calendar/Calendar";
-
-const CalendarScreen = () => {
-  return (
-    <View style={{ flex: 1 }}>
-      <Calendar />
-    </View>
-  );
-}
-
-export default CalendarScreen;
