+++ frontend/src/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import './App.css';
+import MapComponent from './MapComponent';
+
+function App() {
+  return (
+    <div className="App">
+      <MapComponent />
+    </div>
+  );
+}
+
+export default App;
