--- frontend/utils/constants/calendar.js
@@ -1,9 +1,0 @@
-export const MONTH_NAMES = [
-  "Prima", "Duon", "Trine", "Quadra", "Penta", "Hexis", "Septa", "Octus",
-  "Ennea", "Deca", "Primadeca", "Duodeca", "Trideca", "Quadradeca", "Pentadeca",
-  "Nexus"
-];
-export const WEEKDAY_NAMES = [
-  "Primday", "Duday", "Triday", "Quaday", "Penday", "Hexday"
-];
-export const DAYS_PER_MONTH = [23, 21, 21, 22, 21, 21, 22, 21, 21, 22, 21, 21, 23, 21, 21];
