--- frontend/utils/constants/time.js
@@ -1,4 +1,0 @@
-export const SYMBIOND_EPOCH_MILLISECONDS = new Date('2000-01-01T00:00:00.000Z').getTime();
-export const SYMBIOND_HOURS_PER_EARTH_HOUR = 27;
-export const SYMBIOND_HOURS_PER_REGULAR_YEAR = 8694;
-export const SYMBIOND_HOURS_PER_LEAP_YEAR = 9342;
